backend/requirements.txt
# RFP System Backend Dependencies

# Django
Django>=5.2.8
djangorestframework>=3.14.0

# Google Gemini API
google-generativeai>=0.3.0

# Additional utilities
python-dotenv>=1.0.0
requests>=2.31.0
